subo logica pagos y descarga de pdf

+++ BarHub-FrontendAngular/src/app/auth/auth.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { JwtHelperService } from '@auth0/angular-jwt'; 
+import { Router } from '@angular/router';
 
 @Injectable({
   providedIn: 'root'
@@ -9,22 +11,42 @@ export class AuthService {
   private apiUrl = 'http://localhost:8080/api/auth';
   private roles: string[] = [];
   private tokenKey = 'token';
+  private jwtHelper = new JwtHelperService();
 
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient, private router: Router) {}
 
   login(credentials: { username: string; password: string }): Observable<{ token: string }> {
     this.logout();
     return new Observable(observer => {
       this.http.post<{ token: string }>(`${this.apiUrl}/login`, credentials).subscribe({
         next: (res) => {
-          localStorage.setItem(this.tokenKey, res.token); // 👉 guardar token
-          observer.next(res);
+          if (!res.token) { // 👈 Verifica si el token existe
+            observer.error('El servidor no devolvió un token');
+            return;
+          }
+  
+          localStorage.setItem(this.tokenKey, res.token);
+          
+          try {
+            const decodedToken = this.jwtHelper.decodeToken(res.token);
+            
+            localStorage.setItem('user', JSON.stringify({
+              id: decodedToken.id, // ← Campo crítico
+              username: decodedToken.sub
+            }));
+            this.router.navigate(['/carta']);
+
+            observer.next(res);
+          } catch (error) {
+            localStorage.removeItem(this.tokenKey); // Limpiar token inválido
+            observer.error('Token inválido: ' + error);
+          }
           observer.complete();
         },
         error: (err) => observer.error(err)
       });
     });
-  }  
+  }
 
   register(data: { username: string; email: string; password: string }): Observable<any> {
     return this.http.post(`${this.apiUrl}/register`, data);
@@ -57,4 +79,13 @@ export class AuthService {
   setToken(token: string): void {
     localStorage.setItem(this.tokenKey, token);
   }  
+
+  isLoggedIn(): boolean {
+    return localStorage.getItem('user') !== null;
+  }
+
+  getCurrentUser(): any {
+    const user = localStorage.getItem('user');
+    return user ? JSON.parse(user) : null;
+  }
 }
